+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/GitlabLintUtils.kt
@@ -4,15 +4,8 @@ import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import java.nio.file.FileSystems
 
 object GitlabLintUtils {
-    private fun getGitlabLintGlobStrings(): List<String> {
-        val appSettings = AppSettings.instance
-        if (appSettings?.gitlabLintGlobStrings != null) {
-            return appSettings.gitlabLintGlobStrings!!
-        }
-        return listOf("*.gitlab-ci.yml", "*.gitlab-ci.yaml")
-    }
     fun matchesGitlabLintGlob(text: String): Boolean {
-        val globStrings = getGitlabLintGlobStrings()
+        val globStrings = AppSettings.instance.gitlabLintGlobStrings
         val fileSystem = FileSystems.getDefault()
         for (globString in globStrings) {
             val glob = fileSystem.getPathMatcher("glob:$globString")