+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/ui/settings/SettingsConfigurable.kt
@@ -1,82 +1,73 @@
 package com.github.blarc.gitlab.template.lint.plugin.ui.settings
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintBundle.message
+import com.github.blarc.gitlab.template.lint.plugin.extensions.notBlank
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.settings.ProjectSettings
-import com.github.blarc.gitlab.template.lint.plugin.ui.settings.gitlabUrlToken.GitlabUrlToken
+import com.github.blarc.gitlab.template.lint.plugin.ui.settings.gitlabUrlToken.GitlabUrlTokenTable
 import com.intellij.openapi.components.service
-import com.intellij.openapi.options.Configurable
+import com.intellij.openapi.options.BoundConfigurable
 import com.intellij.openapi.project.Project
-import org.jetbrains.annotations.Nls
-import org.jetbrains.annotations.Nullable
-import javax.swing.JComponent
+import com.intellij.ui.ToolbarDecorator
+import com.intellij.ui.dsl.builder.*
 
-class SettingsConfigurable(val project: Project) : Configurable {
-    private var settingsForm: SettingsForm? = null
+class SettingsConfigurable(val project: Project) : BoundConfigurable(message("settings.general.group.title")) {
 
-    // A default constructor with no arguments is required because this implementation
-    // is registered as an applicationConfigurable EP
-    @Nls(capitalization = Nls.Capitalization.Title)
-    override fun getDisplayName(): String {
-        return message("settings.general.group.title")
+    private val gitlabUrlTokenTable = GitlabUrlTokenTable(project)
+    override fun createPanel() = panel {
+        row {
+            comboBox(LintFrequencyEnum.values().toList(), LintFrequencyRenderer())
+                .label(message("settings.frequency.label"))
+                .bindItem(AppSettings.instance::lintFrequency.toNullableProperty())
+        }
+        row {
+            comment(message("settings.frequency.description"))
+        }
+        row {
+            textField()
+                .label(message("settings.remote"))
+                .bindText(project.service<ProjectSettings>()::remote)
+                .validationOnApply { notBlank(it.text) }
+                .align(Align.FILL)
+        }
+        row {
+            checkBox(message("settings.force-https"))
+                .bindSelected(project.service<ProjectSettings>()::forceHttps)
+        }
+        row {
+            checkBox(message("settings.show-merged-preview"))
+                .bindSelected(AppSettings.instance::showMergedPreview)
+                .comment(message("settings.show-merged-preview.comment"))
+        }
+        row {
+            checkBox(message("settings.run-lint-on-file-change"))
+                .bindSelected(AppSettings.instance::runLintOnFileChange)
+                .comment(message("settings.run-lint-on-file-change.comment"))
+        }
+        row {
+            cell(
+                ToolbarDecorator.createDecorator(gitlabUrlTokenTable.table)
+                    .setAddAction { gitlabUrlTokenTable.addGitlabUrlToken() }
+                    .setEditAction { gitlabUrlTokenTable.editGitlabUrlToken() }
+                    .setRemoveAction { gitlabUrlTokenTable.removeGitlabUrlToken() }
+                    .disableUpAction()
+                    .disableDownAction()
+                    .createPanel()
+            ).align(Align.FILL)
+        }.resizableRow()
     }
-
-    @Nullable
-    override fun createComponent(): JComponent {
-        settingsForm = SettingsForm(project)
-        return settingsForm!!.basePanel!!
-    }
-
 
     override fun isModified(): Boolean {
-        val appSettings = AppSettings.instance
-        val projectSettings = project.service<ProjectSettings>()
-
-        if (appSettings != null) {
-            return settingsForm!!.gitlabRemotesTable.isModified(appSettings) ||
-                    settingsForm!!.gitlabUrlTokenTable.isModified(projectSettings) ||
-                    !settingsForm!!.gitlabRemoteTF.contentEquals(projectSettings.remote) ||
-                    settingsForm!!.forceHttpsCB != projectSettings.forceHttps ||
-                    settingsForm!!.lintFrequency != appSettings.lintFrequency ||
-                    settingsForm!!.showMergedPreviewCB != appSettings.showMergedPreview ||
-                    settingsForm!!.runLintOnFileChangeCB != appSettings.runLintOnFileChange
-        }
-        return false
+        return super.isModified() || gitlabUrlTokenTable.isModified()
     }
 
     override fun apply() {
-        val appSettings = AppSettings.instance
-        val projectSettings = project.service<ProjectSettings>()
-
-        if (appSettings != null) {
-            settingsForm!!.gitlabRemotesTable.commit(appSettings)
-            settingsForm!!.gitlabUrlTokenTable.commit(appSettings)
-            projectSettings.remote = settingsForm!!.gitlabRemoteTF!!
-            projectSettings.forceHttps = settingsForm!!.forceHttpsCB
-            appSettings.lintFrequency = settingsForm!!.lintFrequency
-            appSettings.showMergedPreview = settingsForm!!.showMergedPreviewCB
-            appSettings.runLintOnFileChange = settingsForm!!.runLintOnFileChangeCB
-        }
+        gitlabUrlTokenTable.apply()
+        super.apply()
     }
 
     override fun reset() {
-        val appSettings = AppSettings.instance
-        val projectSettings = project.service<ProjectSettings>()
-        val gitlabUrl = projectSettings.gitlabUrl
-
-        if (appSettings != null && gitlabUrl != null) {
-            settingsForm!!.gitlabRemotesTable.tableModel.remotesList = appSettings.remotes.values.toMutableList()
-            settingsForm!!.gitlabUrlTokenTable.tableModel.gitlabUrlTokenList = projectSettings.gitlabUrls
-                .map { GitlabUrlToken(it, appSettings.getGitlabToken(it)) }.toMutableList()
-            settingsForm!!.gitlabRemoteTF = projectSettings.remote
-            settingsForm!!.forceHttpsCB = projectSettings.forceHttps
-            settingsForm!!.lintFrequency = appSettings.lintFrequency
-            settingsForm!!.showMergedPreviewCB = appSettings.showMergedPreview
-            settingsForm!!.runLintOnFileChangeCB = appSettings.runLintOnFileChange
-        }
-    }
-
-    override fun disposeUIResources() {
-        settingsForm = null
+        gitlabUrlTokenTable.reset()
+        super.reset()
     }
 }